Downlad a package zip
