+++ cft/pkg/tmpl/.gitignore
@@ -1,0 +1,1 @@
+modules.zip
