+++ cft/pkg/tmpl/pkg-alias-expect.yaml
@@ -1,0 +1,4 @@
+Resources:
+  foobar:
+    Type: AWS::S3::Bucket
+
