+++ cft/pkg/tmpl/pkg-alias-module.yaml
@@ -1,0 +1,4 @@
+Resources:
+  bar:
+    Type: AWS::S3::Bucket
+
