+++ cft/pkg/tmpl/pkg-alias-template.yaml
@@ -1,0 +1,8 @@
+Rain:
+  Packages:
+    abc:
+      Location: http://localhost:3000/modules.zip
+Resources:
+  foo:
+    Type: $abc/pkg-alias-module.yaml
+
